// src/main/webapp/app/contractmanager/models/apiService.ts
export class ApiService {
  serviceName: string | undefined;
  authorizationEndPoint: string | undefined;
  refreshTokenEndPoint: string | undefined;
  accessTokenEndPoint: string | undefined;
  clientId: string | undefined;
  clientSecret: string | undefined;
  redirectUri: string | undefined;
  tokenScope: string | undefined;
  authorizationCode: string | undefined;
}